New TaskQC class
